creating custom hook for type conversion
